Add a way to view Todos to the page
Adds a dialog form to the page which displays the Todo selected from the
project View section. The necessary event listeners and handlers as well
as a js file, which handles the creation of the dialog form elements,
are added as well. Finally a factory function which creates objects that
hold the relevant information to display for the Todo View is added.