Introduce a way to add Todos using the dialog form
Adds event listeners and handlers to the buttons in the dialog form
which allow the dialog form to be used to add Todos to the projects.
New variables are introduced into the logicModule to keep track of
whether an existing todo is being edited or whether a new todo is being
created. A selectedTodoIndex variable is also added to the logicModule
to keep track of the location of the Todo which was selected. Finally
functions which retrieve the values stored in the form controls of the
dialog form are added to the todoView js file in order to avoid
cluttering the event handlers with that code.