Add the ability to delete projects
Adds buttons to the sidebar project list which allow for a user to
delete the project corresponding to the button pressed. Sidebar.js is
modified to create the buttons used to edit and delete the projects and
an event handler for this process is added in the display controller
IIFE in index.js.