Add the ability to edit project titles
Adds an edit button to projects in the sidebar to allow the user to edit
the title of the project. Also adds the event handler for this processor
in the displayController IIFE in index.js and adds minor styling to
both the project edit and project delete buttons for better
presentation.