Add the ability to delete todos in a project
Adds a button to each todo in the project view todo list to allow a user
to delete the corresponding todo from the project. This introduces
changes to the javascript in projectView which handles the creation of
the elements that hold the todo previews. Some css changes to the
styling for the todo list are also added. Finally the event handler for
this process is added to the displayController.