Add local storage capabilities
Adds local storage to the page. The index.js file is refactored in
addition to the introduction of a storage module which handles the
operations needed for local storage. Both ProjectFactory and TodoFactory
now include a new toJSON function to allow proper serialization of the
objects they make. Finally the webStorage.js file is added to handle
both testing the availability of local storage and to handle retrieval
of storage when available.